+++ src/extensions/yfm/YfmTable/YfmTableSpecs/const.ts
@@ -4,3 +4,9 @@ export enum YfmTableNode {
     Row = 'yfm_tr',
     Cell = 'yfm_td',
 }
+
+export enum YfmTableAttr {
+    Colspan = 'colspan',
+    Rowspan = 'rowspan',
+    CellAlign = 'data-cell-align',
+}